requirements.txt
customtkinter==5.2.2
matplotlib==3.8.0
numpy==1.26.4
Pillow==10.3.0
protobuf==5.27.0
pygame==2.5.2
python-dotenv==1.0.1
Requests~=2.31.0
spotipy==2.23.0
future~=1.0.0
aiohttp~=3.9.5
aiofiles~=22.1.0